Forked project, and added configured DOTNET
