+++ provision/hosts
@@ -51,3 +51,5 @@
 192.168.56.149    	Nanami-Madobe
 192.168.56.150    	SQL_server
 192.168.56.151    	Misaka
+192.168.56.152    	Dot_net
+192.168.56.153    	Mitsuha
