--- provision/lxc/install_chibi_lxc.sh
@@ -1,5 +1,0 @@
-#!/bin/bash
-
-cowsay "instalando chibi lxc"
-pip3 install --upgrade chibi-lxc
-cowsay "termino de instalar chibi lxc"
