+++ provision/lxc/provision/lxc-net
@@ -1,5 +1,5 @@
 USE_LXC_BRIDGE="true"
-LXC_BRIDGE="virbr0"
+LXC_BRIDGE="lxcbr0"
 LXC_ADDR="10.0.3.1"
 LXC_NETMASK="255.255.255.0"
 LXC_NETWORK="10.0.3.0/24"
