--- provision/lxc/provision/lxc/default.conf
@@ -1,7 +1,0 @@
-lxc.idmap = u 0 100000 65536
-lxc.idmap = g 0 100000 65536
-
-lxc.net.0.type = veth
-lxc.net.0.link = virbr0
-lxc.net.0.flags = up
-lxc.net.0.hwaddr = 00:16:3e:xx:xx:xx
