--- provision/lxc/provision/lxc/dnsmasq.conf
@@ -1,1 +1,0 @@
-dhcp-host=playtime,10.0.3.100
