--- provision/lxc/provision/lxc/lxc-usernet
@@ -1,1 +1,0 @@
-vagrant veth virbr0 10
