--- provision/lxc/provision/subgid
@@ -1,2 +1,0 @@
-vagrant:100000:65536
-root:100000:65536
