--- provision/lxc/provision/sysctl.d/00_local_userns.conf
@@ -1,1 +1,0 @@
-user.max_user_namespaces=10000
