+++ provision/nginx/provision/sites_available/waifus.conf
@@ -1,5 +1,7 @@
 # vi: set ft=nginx:
 upstream waifus {
+
+	server mitsuha:80 fail_timeout=1;
 	server misuzu:9200 fail_timeout=1;
 	server pitou:9200 fail_timeout=1;
 	#server ren:9200 fail_timeout=1;
