+++ provision/repos/ganto_lxc3_epel_7.repo
@@ -1,0 +1,10 @@
+[ganto-lxc3]
+name=Copr repo for lxc3 owned by ganto
+baseurl=https://copr-be.cloud.fedoraproject.org/results/ganto/lxc3/epel-7-$basearch/
+type=rpm-md
+skip_if_unavailable=True
+gpgcheck=1
+gpgkey=https://copr-be.cloud.fedoraproject.org/results/ganto/lxc3/pubkey.gpg
+repo_gpgcheck=0
+enabled=1
+enabled_metadata=1
