+++ provision/update_python_lib.sh
@@ -1,6 +1,6 @@
 #!/bin/bash
 set -e
-#set +v
+set +v
 
 echo "update global"
 pip3 install --upgrade pip
@@ -8,7 +8,7 @@ pip3 install --upgrade python-hosts
 pip3 install --upgrade chibi chibi-requests chibi_command chibi-donkey
 
 set +e
-#set -v
+set -v
 
 #cd ~/python_lib/
 #cowsay "iniciando actualizacion de bibliotecas locales de python"
