+++ weeaboo/provision.rb
@@ -20,7 +20,7 @@ class Provision
 			redis: [ Nanoha, Fate ],
 			windows: [ Touko_madobe, Ai_madobe, Yuu_madobe, Nanami_madobe, ],
 			SQL_server: [ Misaka ],
-			LXC: [ Koko ],
+			Dot_net: [Mitsuha],
 		}
 	end
